// packages/client/src/games/splendor/splendorImageIds.ts
export const SPLENDOR_DEV_PUBLIC_IDS: Record<string, string> = {
  "one-white-5": "card-one-white-5_hjvsjq",
  "one-red-2": "card-one-red-2_rkpfyu",
  "one-green-3": "card-one-green-3_ljobkl",
  "one-red-8": "card-one-red-8_vuzxze",
  "one-white-2": "card-one-white-2_oszc44",
  "one-white-4": "card-one-white-4_twby8p",
  "one-black-8": "card-one-black-8_vshh50",
  "one-white-7": "card-one-white-7_cvsxpw",
  "one-black-6": "card-one-black-6_l7yw51",
  "one-green-5": "card-one-green-5_f2uya4",
  "one-red-3": "card-one-red-3_zibqfl",
  "one-blue-5": "card-one-blue-5_xjb0wn",
  "one-green-7": "card-one-green-7_bxlf1a",
  "one-black-7": "card-one-black-7_povpoa",
  "one-red-7": "card-one-red-7_wxl8cl",
  "one-blue-4": "card-one-blue-4_poeu6r",
  "one-blue-1": "card-one-blue-1_ilzlar",
  "one-blue-7": "card-one-blue-7_qkraax",
  "one-green-4": "card-one-green-4_hosah0",
  "one-green-2": "card-one-green-2_tzrzg1",
  "one-red-6": "card-one-red-6_jwyidp",
  "one-red-5": "card-one-red-5_uwtfts",
  "one-green-6": "card-one-green-6_f7kc0n",
  "one-green-1": "card-one-green-1_m9niv8",
  "one-green-8": "card-one-green-8_t0mpd1",
  "one-blue-8": "card-one-blue-8_wsmhvg",
  "one-white-3": "card-one-white-3_ouelq0",
  "one-white-6": "card-one-white-6_tmuugn",
  "one-red-4": "card-one-red-4_noclyn",
  "one-white-1": "card-one-white-1_og4jxi",
  "one-black-5": "card-one-black-5_hcadhh",
  "one-black-4": "card-one-black-4_rsbmiz",
  "one-blue-6": "card-one-blue-6_xqj27y",
  "one-red-1": "card-one-red-1_rchqqd",
  "one-blue-3": "card-one-blue-3_aerna2",
  "one-blue-2": "card-one-blue-2_bvjemt",
  "one-black-3": "card-one-black-3_bduugd",
  "one-black-2": "card-one-black-2_pbdnsx",
  "one-black-1": "card-one-black-1_nbrzhl",
  "one-white-8": "card-one-white-8_fpieyn",
  "two-red-1": "card-two-red-1_h068gk",
  "two-white-3": "card-two-white-3_kfgm4d",
  "two-green-3": "card-two-green-3_gbsiuh",
  "two-green-2": "card-two-green-2_eluu5h",
  "two-blue-3": "card-two-blue-3_v0kcvp",
  "two-white-2": "card-two-white-2_o5ev1w",
  "two-green-1": "card-two-green-1_czvnsi",
  "two-red-4": "card-two-red-4_t0sevb",
  "two-white-1": "card-two-white-1_e06mgr",
  "two-white-5": "card-two-white-5_lacsia",
  "two-green-6": "card-two-green-6_p7vgev",
  "two-blue-2": "card-two-blue-2_gpwqkt",
  "two-blue-5": "card-two-blue-5_wiad76",
  "two-blue-4": "card-two-blue-4_diarro",
  "two-white-4": "card-two-white-4_vtcybp",
  "two-red-6": "card-two-red-6_xnsxdf",
  "two-red-5": "card-two-red-5_shx204",
  "two-red-3": "card-two-red-3_sqsu8t",
  "two-red-2": "card-two-red-2_kcfzdy",
  "two-green-5": "card-two-green-5_aexwgw",
  "two-green-4": "card-two-green-4_szisnf",
  "two-blue-6": "card-two-blue-6_sw14wm",
  "two-blue-1": "card-two-blue-1_ovy8kp",
  "two-black-6": "card-two-black-6_mjz0wv",
  "two-black-5": "card-two-black-5_cwatwq",
  "two-black-4": "card-two-black-4_srmv4i",
  "two-black-3": "card-two-black-3_h51ejg",
  "two-black-2": "card-two-black-2_pjakws",
  "two-black-1": "card-two-black-1_rlb6mr",
  "two-white-6": "card-two-white-6_uvjvk4",
  "three-red-2": "card-three-red-2_f9qdev",
  "three-white-3": "card-three-white-3_qalwix",
  "three-red-4": "card-three-red-4_ek6ubz",
  "three-white-2": "card-three-white-2_lwiwje",
  "three-red-3": "card-three-red-3_ujdrpr",
  "three-green-2": "card-three-green-2_aos1mp",
  "three-green-1": "card-three-green-1_fukkms",
  "three-white-1": "card-three-white-1_q4ed9b",
  "three-green-4": "card-three-green-4_axh5my",
  "three-white-4": "card-three-white-4_jikotv",
  "three-red-1": "card-three-red-1_hey4ma",
  "three-green-3": "card-three-green-3_wcwtjs",
  "three-blue-4": "card-three-blue-4_qkxor2",
  "three-blue-3": "card-three-blue-3_rzhdcx",
  "three-blue-2": "card-three-blue-2_lhpodu",
  "three-blue-1": "card-three-blue-1_j6duw8",
  "three-black-4": "card-three-black-4_mlpm8o",
  "three-black-3": "card-three-black-3_ka0bb0",
  "three-black-2": "card-three-black-2_ovtdwj",
  "three-black-1": "card-three-black-1_l0zral"
} as const;
export const SPLENDOR_NOBLE_PUBLIC_IDS: Record<string, string> = {
  "1": "noble-1_bmc6qq",
  "2": "noble-2_yczysm",
  "3": "noble-3_f4sdxn",
  "4": "noble-4_q4wabd",
  "5": "noble-5_bpkhoz",
  "6": "noble-6_eetgtw",
  "7": "noble-7_nhi4on",
  "8": "noble-8_g5wu0j",
  "9": "noble-9_ffhdvg",
  "10": "noble-10_focrt8"
} as const;
export const SPLENDOR_DECK_BACK_PUBLIC_IDS = {
  "1": "card-one-back-card_rgukzc",
  "2": "card-two-back-card_dnnd8i",
  "3": "card-three-back-card_ryzib9"
} as const;
export const SPLENDOR_NOBLE_BACK_PUBLIC_ID = "noble-back-card_ymlq8e";
export const SPLENDOR_COVER_PUBLIC_ID = "cover_tsasdb";
export const SPLENDOR_CHIP_PUBLIC_IDS = {
  white: "white_fnioda",
  blue: "blue_dzxrxz",
  green: "green_r8hjwe",
  red: "red_a1wxih",
  black: "black_r8wq0z",
  gold: "gold_xadeqd",
} as const;